+++ R/imports.R
@@ -23,4 +23,4 @@ NULL
 # global variables
 utils::globalVariables(c("references", "footnotes", "encoding", "authors", "i",
                          "Name", "Length", "Date", "chunk_number", ".",
-                         "fun_names", "meta_type", "type"))
+                         "fun_names", "meta_type", "type", "full_coverage"))